refactor: combine network capture service and data source
