+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/injection/LogModule.kt
@@ -4,13 +4,11 @@ import io.embrace.android.embracesdk.internal.logs.LogOrchestrator
 import io.embrace.android.embracesdk.internal.logs.LogService
 import io.embrace.android.embracesdk.internal.logs.attachments.AttachmentService
 import io.embrace.android.embracesdk.internal.network.logging.NetworkCaptureDataSource
-import io.embrace.android.embracesdk.internal.network.logging.NetworkCaptureService
 
 /**
  * Holds dependencies that are required for a customer to send log messages to the backend.
  */
 interface LogModule {
-    val networkCaptureService: NetworkCaptureService
     val networkCaptureDataSource: NetworkCaptureDataSource
     val logService: LogService
     val logOrchestrator: LogOrchestrator
