+++ embrace-android-core/src/test/java/io/embrace/android/embracesdk/internal/injection/LogModuleImplTest.kt
@@ -30,7 +30,6 @@ internal class LogModuleImplTest {
             FakeInstrumentationModule(mockk())
         )
 
-        assertNotNull(module.networkCaptureService)
         assertNotNull(module.logService)
         assertNotNull(module.attachmentService)
     }
