+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/injection/FakeLogModule.kt
@@ -3,7 +3,6 @@ package io.embrace.android.embracesdk.fakes.injection
 import io.embrace.android.embracesdk.fakes.FakeConfigService
 import io.embrace.android.embracesdk.fakes.FakeLogOrchestrator
 import io.embrace.android.embracesdk.fakes.FakeNetworkCaptureDataSource
-import io.embrace.android.embracesdk.fakes.FakeNetworkCaptureService
 import io.embrace.android.embracesdk.fakes.FakePayloadStore
 import io.embrace.android.embracesdk.fakes.FakeSessionPropertiesService
 import io.embrace.android.embracesdk.fakes.FakeTelemetryDestination
@@ -13,7 +12,6 @@ import io.embrace.android.embracesdk.internal.logs.LogOrchestrator
 import io.embrace.android.embracesdk.internal.logs.LogService
 import io.embrace.android.embracesdk.internal.logs.attachments.AttachmentService
 import io.embrace.android.embracesdk.internal.network.logging.NetworkCaptureDataSource
-import io.embrace.android.embracesdk.internal.network.logging.NetworkCaptureService
 
 class FakeLogModule(
     override val logService: LogService = EmbraceLogService(
@@ -23,9 +21,6 @@ class FakeLogModule(
         FakePayloadStore(),
     ),
 ) : LogModule {
-
-    override val networkCaptureService: NetworkCaptureService =
-        FakeNetworkCaptureService()
 
     override val logOrchestrator: LogOrchestrator
         get() = FakeLogOrchestrator()
